requirements.txt
feedparser
apscheduler
aniso8601==7.0.0
autoenv==1.0.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.10.3
graphene==2.1.8
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.5
oauthlib==3.1.0
openapi-codec==1.3.2
promise==2.2.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
Rx==1.6.1
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
